requirements.txt
datasets==1.15.1
deep_translator==1.5.4
joblib==1.1.0
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.4
PyYAML==6.0
scikit_learn==1.0.1
torch==1.10.0
tqdm==4.62.3
transformers==4.12.3
